moved getLayoutSetIcon to utils
