+++ frontend/packages/ux-editor/src/utils/getLayoutSetIcon.test.tsx
@@ -2,7 +2,7 @@ import { renderHook } from '@testing-library/react';
 import React from 'react';
 import { getLayoutSetIcon } from './getLayoutSetIcon';
 import type { LayoutSetModel } from 'app-shared/types/api/dto/LayoutSetModel';
-import { QuestionmarkIcon } from '@studio/icons';
+import { QuestionmarkIcon } from 'libs/studio-icons/src';
 
 describe('useLayoutSetIcon', () => {
   it('should return default icon for unknown types', () => {
